Backend/requirements.txt
asgiref==3.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.16
django-cors-headers==4.3.1
django-createsuperuser==2020.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-prometheus==2.3.1
dnspython==2.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pillow==10.2.0
prometheus_client==0.24.1
PyJWT==2.4.0
python-dotenv==1.2.1
psycopg2-binary==2.9.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
setuptools==80.9.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0